refactor: migrate insights components to a flat directory structure and update module references
