+++ src/ravioli/frontend/src/components/Insights.ts
